requirements.txt
chalice
dnspython
Jinja2
pymongo
python-slugify
